Add additional features to the user profile page
I changed the major field from an Input to a Select. In order to
do this, I had to add a new Select tag, and wanting it to work
seamlessly with the existing structure used by Input, I added a new
handler for Select objects to `form-hooks.js` and used that handler
when creating a Select tag. To facilitate creating a Select tag based
upon majors, I created a Javascript file under Utility/ which contains
all valid majors in an array. I also added a new SuccessModal to
differentiate between error messages and benign messages.

I modified the validator.js methods to not do any string comparisons
and terminate when any invalid validator is found, and this helped the
latency issues with validation of many inputs significantly; however,
it does still appear sluggish upon startup, and seems to take about
5 seconds to give feedback about improper inputs, and I'm not sure
why.

I tied the submission button to the existing `hadChanges` function,
and this works great. I'm not sure how we'd use validity alone,
because it's entirely possible for a field to be valid but not ready
to submit. For example, it's valid if the major dropdown contains the
user's current major, but we still can't allow submission unless the
value is different. For what it's worth, `hadChanges` does rely on
`isValid` significantly and doesn't redo any work done by the
validation.